test: cover text object prefixes
